\docType{data}
\name{elementCodes}
\alias{elementCodes}
\title{NRCS Element Codes}
\format{A data frame with 102 rows and 3 variables}
\usage{
data(elementCodes)
}
\description{
A dataset containing Element names (eg. "Air Temperature Max") and the corresponding abbreviated element code ("TMAX).
}
\keyword{datasets}

