% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabAgriMetData.R
\name{getAgriMet.data}
\alias{getAgriMet.data}
\title{grabNRCS.data}
\usage{
getAgriMet.data(site_id, timescale, DayBgn, DayEnd, pCodes)
}
\arguments{
\item{site_id}{An AgriMet station code (4 upper-case characters). A station map with codes is located here: \url{https://www.usbr.gov/pn/agrimet/agrimetmap/agrimap.html}.\cr}

\item{timescale}{Specify the desired timescale of the data. Valid entries are 'subhourly', hourly', or 'daily'.\cr}

\item{DayBgn}{Specify the beginning date (as YYYY-MM-DD ) for the returned data.\cr}

\item{DayEnd}{Specify the end date (as YYYY-MM-DD ) for the returned data.\cr}

\item{pCodes}{A character vector of AgriMet parameter codes you'd like returned. Codes and more information can be found on the AgriMet website: \url{https://www.usbr.gov/pn/agrimet/aginfo/station_params.html}}
}
\value{
Returns a data frame of requested data. Only elements with at least one data will be returned in the data frame.\cr
}
\description{
A data downloading tool for NRCS networks. Data retrieval is limited by the speed of the connection, if timeout errors persist break requested time periods down into smaller chunks, or use a faster connection.
}
\examples{
\dontrun{
# Return wind data at the Nampa Station, for Jan 2020
getAgriMet.data(site_id="NMPI", timescale="subhourly", DayBgn = "2020-01-01", DayEnd="2020-02-01", pCodes=c("WD", "WG", "WS"))
}
}
\references{
Downloads <https://wcc.sc.egov.usda.gov/reportGenerator>
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}
}
\keyword{atmopsheric}
\keyword{atmosphere,}
\keyword{climate,}
\keyword{data,}
\keyword{environment,}
\keyword{environmental}
\keyword{in-situ,}
\keyword{weather\cr}
