% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabNRCSelements.R
\name{grabNRCS.elements}
\alias{grabNRCS.elements}
\title{grabNRCS.elements}
\usage{
grabNRCS.elements(site_id = "SCAN:2221")
}
\arguments{
\item{site_id}{site id(s) of NRCS site(s) to grab metadata for.  Default is 'SCAN:2221'.  The user should note that it may take several minutes to download element level data for an individual NRCS network, e.g., SCAN.}
}
\value{
A list of n dataframes comprising the element level metadata for n NRCS sites
}
\description{
A search tool that grabs element metadata (reporting variables) for sites within networks managed by the NRCS. Element level metadata can be pulled for the following networks: i) SCAN, ii) SNOTEL (SNTL), iii) SNOTEL-Lite (SNTLT), or iv) OTHER\cr
}
\examples{
grabNRCS.elements(site_id='SCAN:2221',verbose=FALSE)
grabNRCS.elements(site_id=c('SCAN:2221','SNTL:1277','SNTL:1252'))
\\cr
#You can make use of both 'grab' functions, \\code{grabNRCS.meta()} & \\code{grabNRCS.elements()} to grab site and element level data for a specific network
SNTL.sites<-grabNRCS.meta(ntwrks="SNTL")
grabNRCS.elements(site_id=SNTL.sites$SNTL$site_id)
#It may take up to several minutes to download element level data for an entire NRCS network
}
\author{
Josh Roberti \email{jaroberti87@gmail.com}
}
\references{
element level information can be found here: https://www.wcc.nrcs.usda.gov/web_service/AWDB_Web_Service_Reference.htm#commonlyUsedElementCodes
}
\seealso{
Currently none
}
\keyword{atmopsheric}
\keyword{atmosphere,}
\keyword{climate,}
\keyword{data,}
\keyword{environment,}
\keyword{environmental}
\keyword{in-situ,}
\keyword{weather}

