% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabNRCSmeta.R
\name{grabNRCS.meta}
\alias{grabNRCS.meta}
\title{grabNRCS.meta}
\usage{
grabNRCS.meta(ntwrks = "SCAN", cnvrt.elev = FALSE)
}
\arguments{
\item{ntwrks}{Name(s) of NRCS network(s) to grab metadata for.  Options are: 'SCAN','SNTL','SNTLT','SNOW','MPRC','OTHER','COOP','USGS','MSNT','BOR','CLMIND' and/or 'ALL'. If the user enters the term 'ALL,' site level metadata for all networks will be returned.  Default is 'SCAN'\cr}

\item{cnvrt.elev}{Optional. If set to TRUE, the elevation data will be converted from feet to meters.  Default is set to FALSE, i.e., elevation data are output in units of feet.}
}
\value{
A list of n dataframes comprising the site level metadata for n NRCS networks\cr
}
\description{
A search tool that grabs site level metadata for NRCS networks: SCAN, SNOTEL, SNOTEL-LITE, SNOW, MPRC, and OTHER
}
\examples{
grabNRCS.meta(ntwrks="SCAN")
}
\references{
Downloads <https://www.wcc.nrcs.usda.gov/web_service/AWDB_Web_Service_Reference.htm>
}
\seealso{
Currently none
}
\author{
Josh Roberti \email{jaroberti87@gmail.com}
}
\keyword{atmopsheric}
\keyword{atmosphere,}
\keyword{climate,}
\keyword{data,}
\keyword{environment,}
\keyword{environmental}
\keyword{in-situ,}
\keyword{weather\cr}
