% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grabNRCSdata.R
\name{grabNRCS.data}
\alias{grabNRCS.data}
\title{grabNRCS.data}
\usage{
grabNRCS.data(network, site_id, timescale, DayBgn, DayEnd)
}
\arguments{
\item{network}{The network of the NRCS/AWDB site of interest. Currently only works for options: 'SCAN','SNTL','SNTLT', and 'OTHER'.\cr}

\item{site_id}{The NRCS site ID. Use grabNRCS.meta to retrieve a list of available sites in a specified network. \cr}

\item{timescale}{Specify the desired timescale of the data. Typically 'hourly', 'daily', or 'monthly'\cr}

\item{DayBgn}{Optional. Specify the beginning date for the returned date, otherwise the beginning of the period of record is returned.\cr}

\item{DayEnd}{Optional. Specify the end date for the returned date, otherwise the end of the period of record is returned.\cr}
}
\value{
Returns a data frame of requested data. If the output is not specified, the data frame is returned as 'NRCS.df' to the global environment.\cr
}
\description{
A data downloading tool for NRCS networks.
}
\examples{
Currently none
}
\references{
A list networks and their identifiers can be found here: https://www.wcc.nrcs.usda.gov/report_generator/AWDB_Network_Codes.pdf
}
\seealso{
Currently none
}
\author{
Robert Lee \email{rhlee@colorado.edu}
}
\keyword{atmopsheric}
\keyword{atmosphere,}
\keyword{climate,}
\keyword{data,}
\keyword{environment,}
\keyword{environmental}
\keyword{in-situ,}
\keyword{weather\cr}

